+++ molo/core/tests/test_models.py
@@ -7,7 +7,7 @@ from django.utils import timezone
 from django.core.urlresolvers import reverse
 from django.contrib.auth.models import User
 
-from molo.core.models import ArticlePage, SiteLanguage
+from molo.core.models import ArticlePage, SiteLanguage, PageTranslation
 from molo.core import constants
 from molo.core.templatetags.core_tags import (
     load_descendant_articles_for_section)
@@ -306,3 +306,22 @@ class TestModels(TestCase, MoloTestCaseMixin):
         self.assertContains(response, 'English')
         self.assertContains(response, 'français')
         self.assertNotContains(response, 'español')
+
+    def test_signal_on_page_delete_removes_translations(self):
+        section = self.mk_section(
+            self.section_index, title="Section", slug="section")
+
+        page1a, page2a = self.mk_articles(section, 2)
+        self.mk_article_translation(page1a, self.french)
+        page2b = self.mk_article_translation(page2a, self.french)
+        translation2ab = PageTranslation.objects.get(translated_page=page2b)
+
+        page1a.delete()
+
+        self.assertEqual(
+            list(ArticlePage.objects.all().values('id')),
+            [{'id': page2a.id}, {'id': page2b.id}])
+
+        self.assertEqual(
+            list(PageTranslation.objects.all().values('id')),
+            [{'id': translation2ab.id}])