+++ CHANGES.rst
@@ -1,6 +1,12 @@
 CHANGE LOG
 ==========
 
+3.2.1
+------
+
+- Delete translated page when a page is deleted
+- Added extra lang info for languages that django doesn't support
+
 3.2.0
 ------